requirements.txt
Sanic==0.8.3
sanic-openapi
requests-file
jsonschema
aiokafka==0.4.0
apache-libcloud
trio
kafka
shapely
boto3
